fix(compat): dinput.h stub + GetDoubleClickTime/CreateDirectory/SHGetSpecialFolderPath stubs
Add compat/dinput.h with all standard DIK_* keyboard scan-code constants
so KeyDefs.h can compile without the real DirectX SDK on non-Windows.

Add to compat/windows.h:
  - GetDoubleClickTime() — returns 500ms default
  - CreateDirectoryA() — backed by POSIX mkdir()

Add to compat/shlobj.h:
  - CSIDL_* special folder constants
  - SHGetSpecialFolderPathA() — returns $HOME for all folder IDs

Fix INI.cpp:625: (UnsignedInt)userData → (UnsignedInt)(uintptr_t)userData
to avoid pointer-truncation warning on 64-bit Clang/GCC.

+++ GeneralsMD/Code/GameEngine/Source/Common/INI/INI.cpp
@@ -622,7 +622,7 @@ void INI::parseBool( INI* ini, void * /*instance*/, void *store, const void* /*u
 void INI::parseBitInInt32( INI *ini, void *instance, void *store, const void* userData )
 {
 	UnsignedInt* s = (UnsignedInt*)store;
-	UnsignedInt mask = (UnsignedInt)userData;
+	UnsignedInt mask = (UnsignedInt)(uintptr_t)userData;  // userData holds a bit-mask, not a pointer
 
 	if (INI::scanBool(ini->getNextToken()))
 		*s |= mask;
